Update sur les controller et ajout d'exception custom
ajout de middleware et du README

+++ IBay/IBay/Controllers/ProductController.cs
@@ -2,7 +2,9 @@
 using DAL.Model;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using static IBay.AuthorizationMiddleware;
 using Swashbuckle.AspNetCore.Annotations;
+using System.Security.Claims;
 
 namespace IBay.Controllers
 {
@@ -15,12 +17,21 @@ namespace IBay.Controllers
         [HttpPost("{sellerId:int}")]
         [SwaggerOperation("Create a new product")]
         [SwaggerResponse(200, "Product created successfully")]
-        public IActionResult Create(int sellerId, string productName, string productDescription,
-            ProductType productType, double productPrice, int productStock)
+        public IActionResult Create(int sellerId, string productName, string productDescription,ProductType productType, double productPrice, int productStock)
         {
-            var newProduct = context.CreateProduct(sellerId, productName, productDescription,
+            try
+            {
+                var newProduct = context.CreateProduct(sellerId, productName, productDescription,
                 productType, productPrice, productStock);
-            return Ok(newProduct);
+                return Ok(newProduct);
+            }
+            catch(IbayContext.NotFoundException ex){
+                return NotFound(ex.Message);
+            }
+            catch (IbayContext.BadRequestException ex) {
+                return BadRequest(ex.Message);
+            }
+            
         }
 
         [HttpGet]
@@ -29,15 +40,6 @@ namespace IBay.Controllers
         public IActionResult Get()
         {
             var products = context.GetProducts();
-            return Ok(products);
-        }
-        
-        [HttpGet("/sort")]
-        [SwaggerOperation("Get sorted products")]
-        [SwaggerResponse(200, "List of sorted products")]
-        public IActionResult Get(SortCategory sortCategory, int limit)
-        {
-            var products = context.GetProductSortedBy(sortCategory, limit);
             return Ok(products);
         }
 
@@ -55,19 +57,56 @@ namespace IBay.Controllers
             return Ok(product);
         }
 
+        [HttpGet("sort")]
+        [SwaggerOperation("Get sorted products")]
+        [SwaggerResponse(200, "List of sorted products")]
+        public IActionResult Get(SortCategory sortCategory, int limit)
+        {
+            try
+            {
+                var products = context.GetProductSortedBy(sortCategory, limit);
+                return Ok(products);
+            }
+            catch(IbayContext.BadRequestException ex)
+            {
+                return BadRequest(ex.Message);
+            }
+            
+        }
+
         [HttpPut("{id:int}")]
         [SwaggerOperation("Update a product")]
         [SwaggerResponse(200, "Product updated successfully")]
         [SwaggerResponse(404, "Product not found")]
-        public IActionResult Update(int id, [FromBody] Product product)
+        public IActionResult Update(int id, string updatedProductName, string updatedProductDescription, ProductType updatedProductType, double updatedProductPrice, int updatedProductStock, bool updatedProductAvailable )
         {
-            var updatedProduct = context.UpdateProduct(id, product.ProductName, product.ProductDescription,
-                product.ProductType, product.ProductPrice, product.ProductStock, product.Available);
-            if (updatedProduct == null)
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            var productToCheck = context.GetProductById(id);
+
+            if (userId == null || productToCheck == null)
             {
-                return NotFound();
+                return Forbid();
             }
-            return Ok(updatedProduct);
+            if(productToCheck.Seller.UserId != int.Parse(userId))
+            {
+                Console.WriteLine("Not product owner");
+                return Forbid();
+            }
+
+            try
+            {
+                var updatedProduct = context.UpdateProduct(id, updatedProductName, updatedProductDescription, updatedProductType, updatedProductPrice, updatedProductStock, updatedProductAvailable);
+                if (updatedProduct == null)
+                {
+                    return NotFound();
+                }
+                return Ok(updatedProduct);
+            }
+            catch(IbayContext.NotFoundException ex)
+            {
+                return NotFound(ex.Message);
+            }
+            
         }
 
         [HttpDelete("{id:int}")]
@@ -76,12 +115,33 @@ namespace IBay.Controllers
         [SwaggerResponse(404, "Product not found")]
         public IActionResult Delete(int id)
         {
-            var product = context.DeleteProduct(id);
-            if (product == null)
+            var userId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            var productToCheck = context.GetProductById(id);
+
+            if (userId == null || productToCheck == null)
             {
-                return NotFound();
+                return Forbid();
             }
-            return Ok(product);
+            if (productToCheck.Seller.UserId != int.Parse(userId))
+            {
+                Console.WriteLine("Not product owner");
+                return Forbid();
+            }
+
+            try
+            {
+                var product = context.DeleteProduct(id);
+                if (product == null)
+                {
+                    return NotFound();
+                }
+                return Ok(product);
+            }
+            catch(IbayContext.NotFoundException ex)
+            {
+                return NotFound(ex.Message);
+            }
+            
         }
     }
 }
